pdf-processor/requirements.txt
# Existing dependencies
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
pdfminer.six==20221105
spacy==3.6.1
regex==2023.8.8

# gRPC dependencies
grpcio>=1.71.0
grpcio-tools>=1.71.0
protobuf>=4.24.4